datatype van lat/lang decimal naar float veranderd, get driverLocation methode geÃ¯mplementeerdin userController, 6 nieuwe drivers aangemmakt met Location in DBInitializer, ook bij customers en oude drivers location toegevoegd

+++ WebApi/WebApi/WebApi/Models/Users/Location.cs
@@ -8,7 +8,7 @@ namespace WebApi.Models.Users
     public class Location
     {
         public int LocationID { get; set; }
-        public decimal latitude { get; set; }
-        public decimal longitude { get; set; }
+        public float latitude { get; set; }
+        public float longitude { get; set; }
     }
 }
